#32 Refactor To Concurrency and MVVM For Async Calls

+++ ArgosIOS/Argos/Components/ThermometerView.swift
@@ -9,7 +9,7 @@ import SwiftUI
 
 struct ThermometerView: View {
     
-    @State var current: Float
+    @State var current: Double
     let minimum: Float
     let maximum: Float
     let label: String
@@ -18,7 +18,7 @@ struct ThermometerView: View {
         VStack(spacing: 20) {
             ArgosHeader(self.label)
             
-            Gauge(value: self.current, in: self.minimum...self.maximum) {
+            Gauge(value: Float(self.current), in: self.minimum...self.maximum) {
                 Image(systemName: "thermometer")
                     .font(.caption)
             } currentValueLabel: {
